# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Sample API",
    "description": "API description in Markdown.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "Hello"
        ],
        "summary": "Hello Word.",
        "description": "Optional extended description in Markdown.",
        "produces": [
          "text/html"
        ],
        "responses": {
          "200": {
            "description": "Hello word",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/items": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Get ToDo items",
        "description": "Returns list of todo items",
        "responses": {
          "200": {
            "description": "List of items",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ToDoItem"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Items"
        ],
        "summary": "Create ToDo Item",
        "parameters": [
          {
            "in": "body",
            "name": "ToDo",
            "schema": {
              "$ref": "#/definitions/ToDoItem"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/items/{toDoId}": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Get single ToDo",
        "parameters": [
          {
            "in": "path",
            "name": "toDoId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "To Do Item",
            "schema": {
              "$ref": "#/definitions/ToDoItem"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Items"
        ],
        "summary": "Update selected ToDo",
        "parameters": [
          {
            "in": "path",
            "name": "toDoId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Items"
        ],
        "summary": "Remove item ToDo",
        "parameters": [
          {
            "in": "path",
            "name": "toDoId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "ToDoItem": {
      "properties": {
        "_id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "completed": {
          "type": "boolean"
        }
      },
      "required": [
        "title"
      ]
    }
  }
}